test(google): move `google_cached_content` tests to `tests/models/google/test_cache.py`
