+++ tests/models/google/cassettes/test_cache/test_google_model_cached_content.yaml
